made changes discussed with kevin.
removed the entangled flows
cleaned up the way I was handling the BottomNavBarDestinations
and bottomBar.

added the ChooseUTopBar to all the main destinations individually.

removed unnecessary callback in the MainScreenRoutes

removed the onBackPress callback from the super class of NavigationManger.

fixed the way I was handling the selectedTab in the BottomNavViewModel

changed the name of MainViewModel to BottomNavViewModel

removed unnecessary Paramter in the ToolBarState.Home  I didn't need both hideTrailingIcon and showNotification Icon since they are in the same spot.